+++ core-ng/src/main/java/core/framework/internal/async/ExecutorTask.java
@@ -12,40 +12,36 @@ import javax.annotation.Nullable;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.List;
+import java.util.Set;
 import java.util.concurrent.Callable;
 
 /**
  * @author neo
  */
-public class ExecutorTask<T> implements Callable<T> {
+class ExecutorTask<T> implements Callable<T> {
     private static final Logger LOGGER = LoggerFactory.getLogger(ExecutorTask.class);
-    final String actionId;
-    private final String action;
-    private final LogManager logManager;
+
     private final Callable<T> task;
-    private final Instant startTime;
-    private final long maxProcessTimeInNano;
+    private final LogManager logManager;
+    private final TaskContext context;
+
     private final String rootAction;
-    private final String refId;
     private final String correlationId;
+    private final String refId;
     private final Trace trace;
     @Nullable
     private final PerformanceWarning[] warnings;
 
-    ExecutorTask(Callable<T> task, LogManager logManager, TaskContext context) {
+    ExecutorTask(Callable<T> task, LogManager logManager, TaskContext context, ActionLog parentActionLog) {
         this.task = task;
         this.logManager = logManager;
-        actionId = context.actionId;
-        action = context.action;
-        startTime = context.startTime;
-        maxProcessTimeInNano = context.maxProcessTimeInNano;
-        ActionLog parentActionLog = context.parentActionLog;
+        this.context = context;
         if (parentActionLog != null) {  // only keep info needed by call(), so parentActionLog can be GCed sooner
             List<String> parentActionContext = parentActionLog.context.get("root_action");
-            rootAction = parentActionContext != null ? parentActionContext.get(0) : parentActionLog.action;
+            rootAction = parentActionContext != null ? parentActionContext.getFirst() : parentActionLog.action;
             correlationId = parentActionLog.correlationId();
             refId = parentActionLog.id;
-            trace = parentActionLog.trace;
+            trace = parentActionLog.trace == Trace.CASCADE ? Trace.CASCADE : null;  // trace only with parent.cascade
             warnings = parentActionLog.warnings();
         } else {
             rootAction = null;
@@ -59,50 +55,46 @@ public class ExecutorTask<T> implements Callable<T> {
     @Override
     public T call() throws Exception {
         VirtualThread.COUNT.increase();
-        ActionLog actionLog = logManager.begin("=== task execution begin ===", actionId);
+        ActionLog actionLog = logManager.begin("=== task execution begin ===", context.actionId);
         try {
             actionLog.action(action());
-            actionLog.warningContext.maxProcessTimeInNano(maxProcessTimeInNano);
-            // here it doesn't log task class, is due to task usually is lambda or method reference, it's expensive to inspect, refer to ControllerInspector
+            actionLog.warningContext.maxProcessTimeInNano(context.maxProcessTimeInNano);
+            // here doesn't log task class, due to task usually is lambda or method reference, it's expensive to inspect, refer to ControllerInspector
             if (rootAction != null) { // if rootAction != null, then all parent info are available
                 actionLog.context("root_action", rootAction);
                 LOGGER.debug("correlationId={}", correlationId);
                 actionLog.correlationIds = List.of(correlationId);
                 LOGGER.debug("refId={}", refId);
                 actionLog.refIds = List.of(refId);
-                if (trace == Trace.CASCADE) actionLog.trace = Trace.CASCADE;
+                if (trace != null) actionLog.trace = trace;
                 if (warnings != null) actionLog.initializeWarnings(warnings);
             }
             LOGGER.debug("taskClass={}", CallableTask.taskClass(task).getName());
-            Duration delay = Duration.between(startTime, actionLog.date);
+            Duration delay = Duration.between(context.startTime, actionLog.date);
             LOGGER.debug("taskDelay={}", delay);
             actionLog.stats.put("task_delay", (double) delay.toNanos());
             actionLog.context.put("thread", List.of(Thread.currentThread().getName()));
             return task.call();
         } catch (Throwable e) {
             logManager.logError(e);
-            throw new TaskException(Strings.format("task failed, action={}, id={}, error={}", action, actionId, e.getMessage()), e);
+            throw new TaskException(Strings.format("task failed, action={}, id={}, error={}", context.action, context.actionId, e.getMessage()), e);
         } finally {
             logManager.end("=== task execution end ===");
+            context.runningTasks.remove(toString());
             VirtualThread.COUNT.decrease();
         }
     }
 
     String action() {
-        return rootAction == null ? "task:" + action : rootAction + ":task:" + action;
+        return rootAction == null ? "task:" + context.action : rootAction + ":task:" + context.action;
     }
 
-    // used to print all canceled tasks during shutdown
+    // used to print all canceled tasks during shutdown, used by both ScheduledExecutorService and VirtualThreadExecutor
     @Override
     public String toString() {
-        return action() + ":" + actionId;
+        return action() + ":" + context.actionId;
     }
 
-    static class TaskContext {
-        String actionId;
-        String action;
-        Instant startTime;
-        ActionLog parentActionLog;
-        long maxProcessTimeInNano;
+    record TaskContext(String actionId, String action, Instant startTime, long maxProcessTimeInNano, Set<String> runningTasks) {
     }
 }