+++ core-ng/src/test/java/core/framework/internal/async/ExecutorTaskTest.java
@@ -6,6 +6,7 @@ import core.framework.log.ActionLogContext;
 import org.junit.jupiter.api.Test;
 
 import java.time.Instant;
+import java.util.HashSet;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -16,16 +17,16 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 class ExecutorTaskTest {
     @Test
     void action() {
-        assertThat(new ExecutorTask<Void>(() -> null, null, context(null)).action())
+        assertThat(new ExecutorTask<Void>(() -> null, null, context(), null).action())
             .isEqualTo("task:action");
 
         var parentActionLog = new ActionLog(null, null);
         parentActionLog.action = "parentAction";
-        assertThat(new ExecutorTask<Void>(() -> null, null, context(parentActionLog)).action())
+        assertThat(new ExecutorTask<Void>(() -> null, null, context(), parentActionLog).action())
             .isEqualTo("parentAction:task:action");
 
         parentActionLog.context("root_action", "rootAction");
-        assertThat(new ExecutorTask<Void>(() -> null, null, context(parentActionLog)).action())
+        assertThat(new ExecutorTask<Void>(() -> null, null, context(), parentActionLog).action())
             .isEqualTo("rootAction:task:action");
     }
 
@@ -33,7 +34,7 @@ class ExecutorTaskTest {
     void callWithException() {
         var task = new ExecutorTask<Void>(() -> {
             throw new Error("test");
-        }, new LogManager(), context(null));
+        }, new LogManager(), context(), null);
         assertThatThrownBy(task::call)
             .isInstanceOf(TaskException.class)
             .hasMessageContaining("task failed")
@@ -46,23 +47,17 @@ class ExecutorTaskTest {
         var task = new ExecutorTask<>(() -> {
             assertThat(ActionLogContext.get("thread")).hasSize(1);
             return Boolean.TRUE;
-        }, new LogManager(), context(null));
+        }, new LogManager(), context(), null);
         assertThat(task.call()).isTrue();
     }
 
     @Test
     void convertToString() {
-        assertThat(new ExecutorTask<Void>(() -> null, null, context(null)).toString())
+        assertThat(new ExecutorTask<Void>(() -> null, null, context(), null).toString())
             .isEqualTo("task:action:actionId");
     }
 
-    private ExecutorTask.TaskContext context(ActionLog parentActionLog) {
-        var context = new ExecutorTask.TaskContext();
-        context.actionId = "actionId";
-        context.action = "action";
-        context.startTime = Instant.now();
-        context.parentActionLog = parentActionLog;
-        context.maxProcessTimeInNano = 25_000_000_000L;
-        return context;
+    private ExecutorTask.TaskContext context() {
+        return new ExecutorTask.TaskContext("actionId", "action", Instant.now(), 25_000_000_000L, new HashSet<>());
     }
 }